+++ Atum/Support/MetaData/FloatWidget.cpp
@@ -30,6 +30,6 @@ void FloatWidget::SetData(void* set_data)
 
 void FloatWidget::SetEditedData()
 {
-	*data = atof(ebox->GetText());
+	*data = (float)atof(ebox->GetText());
 	changed = true;
 }
